tfcli: reorganize files as a single package and file-per-struct
process: remove unused process package

client: make workspace store map safe for concurrent use

client: make operation functions safe for concurrent calls

client: clean up the terraform lock file if it leaked
