+++ pkg/client/templates/main.tf.json.tpl
