+++ pkg/pipeline/controller.go
@@ -21,6 +21,8 @@ import (
 	"path/filepath"
 	"strings"
 
+	"github.com/crossplane-contrib/terrajet/pkg/resource"
+
 	"github.com/muvaf/typewriter/pkg/wrapper"
 	"github.com/pkg/errors"
