+++ pkg/pipeline/terraformed.go
@@ -23,6 +23,8 @@ import (
 	"path/filepath"
 	"strings"
 
+	"github.com/crossplane-contrib/terrajet/pkg/resource"
+
 	"github.com/muvaf/typewriter/pkg/wrapper"
 	"github.com/pkg/errors"
