+++ pkg/resource/interfaces.go
